equiv: don't mess with equiv in faked environments
The equiv lists are problematic and this is not the correct solution.  This
is a hack that improves the situation a little.

The whole equiv code doesn't work.  They're stored in the sm_state and so
they shouldn't be modified after the state is set but they are...  The
worst part is the remove_from_equiv() and you can end up in a situation
where a variable is not equivalent to itself which causes a lot of
problems.

I wanted the equiv code to work fast, so I can't just use the compare code.
I'm not really sure what to do.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>