free_strict: take a short cut in unmatched_state() for non-parameters
The code in unmatched_state() is supposed to handle something like this:

void my_free(void *p)
{
	if (p)
		free(p);
}

In the unmatched state path "p" is NULL so let's just say it is free
instead of saying that the it's &undefined.

But there isn't any need to do that for situations where the pointer "p"
is not a parameter.

What worries me about this commit is that I don't remember working on this
so although this code seems clear enough and self explanatory, I secretly
worry that I was actually trying to fix a bug?

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>