extra: fix a type issue handling unknown for loops
These loops look like: "for (i = 0; i < unknown_variable; i++) {".
Originally, Smatch would say that "i" was in the 0-s32max range but if
the unknown_variable is a u16 then the range is actually "0-u16max - 1".

What this did is that at the end of loop if there is an exit check which
said that:

	if (i == unknown_variable) {

Then Smatch would say that's impossible because s32max is higher than
u16max so it wouldn't calculate the implications correctly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>