comparison: fix a bug dealing with transitive comparisons
Say we know that "a < c" and "c < b" then we know that "a < c".  The
problem is that Smatch always stores these relationships in alphabetical
order "a vs b" not "b vs a".  So in the loop to copy these it changes which
is the left variable.  We need to reset it back to the start with each
iteration of the loop or it starts comparing the wrong variables.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>