extra: preserve some sm states dealing with PARAM_LIMIT
This is a partial solution but at least it's technically accurate.  The
problem is that when we dereference a pointer then Smatch say "Okay, this
pointer is non-NULL but it's not an assignment." so in some instances that
is treated as a PARAM_LIMIT test.

Anyway, say we don't know anything about a parameter going into a function.
And then at the end we know it's either zero, valid or an error pointer.
Even though that means we have eliminated the values of 1-4095, that's
still not really useful information.  So it's better to just keep recorded
as unknown if we have a sm state with history associated with the old
state.

Historically, if the PARAM_LIMIT was 0 or valid then that was treated as
the same as unknown as well, but it's not because the function may have
tested for error pointers explicitly.  This is fundamentally a problem
that's always going to happen if we treat dereferencing the same as a
PARAM_LIMIT and I don't know how to address that.

But for now this change is very conservative and shouldn't break anything.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>