states,implied: delete __set_sm_fake_stree()
I cannot think of a reason for the __set_sm_fake_stree() function to
exist.  I think it is a relic of the nastiness before I fixed how fake
strees work.

I recently had to debug a bug in implications and I worried about this
code but if I delete it then that makes everything more uniform.  I have
built a kernel with this and didn't notice any negative fallout.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>