scripts: add xen script
This script can be used to detect all candidates for speculative leak
gadgets. This script only takes care of building the hypervisor with the
smatch tool support. As long as new defect candidates are detected, the
script will continue. Furthermore, an upper bound is set to stop the
detection at some point. Note, smatch is tracking state locally, and
miproves its results when being executed multiple times in a row.

To detect newly introduced defects, this tool should be used on the old
change and the new change, so that the differences in the output can be
compared. Due to the time based behavior of smatch, there might be
differences in the output that are not caused by code changes.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>