fix null-pointer crash with with ident same as one of the attributes
match_attribute() will crash when the token has the same identifier
as one of the attributes but is not an attribute. In this case,
the corresponding symbol_op will be null but this is not checked.
This seems to happen only with old-style declarations.

Fix this by adding the missing null-check.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>