check_list_add: warn about use after frees with lists
This is a simple/limited check for a certain class of use after free.
The code allocates a struct, adds it to a list, then encounters an
error and frees the struct without removing it from the list.

This check has some false positives where removing the struct from the
list happens in another thread.  Less commonly where removing the struct
happens in a put_device().  I hopefully will eventually make parsing
put_device() work better.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>