no needs to use MARK_CURRENT_DELETED() for multi-jumps
MARK_CURRENT_DELETED() was added for the case(s) where an element
must be removed from the list but the address of the other elements
must not be changed. In this case of effectively removing the
element from it list, the element is 'marked' as deleted in the list
and the list walking macros will later take this in account.

However, this is never needed for multi-jumps.

So, use the usual DELETE_CURRENT_PTR() for them.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>