scheck: predefine __SYMBOLIC_CHECKER__
It can be useful to use the same testcase for the symbolic checker
and normal sparse (or test-linearize) processing.

So, there must be a mean to somehow ignore the assertions used
for the symbolic checker when it's not used (since these are otherwise
not known to sparse).

Resolve this by adding a predefine, __SYMBOLIC_CHECKER__, to the
symbolic checker.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>