function_hooks: fake parameter assignments
This is to handle function which do something like:

int my_value(struct foo *p)
{
	return p->value;
}

If the caller does: "x = my_value(p);" then we want that to create a fake
assignment "x = p->value;".  The trick is that a lot of code does stuff
like "foo = strip_expr(foo);" where "foo" is on both sides of the
assignment so we have to generate the fake assignment, store it, and then
parse it at the end.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>