type: introduce get_comparison_type()
If you have a comparison, the type is not really clear.  There is the
real C type which is &int_ctype and then there is type that Smatch returns
for get_type().  The Smatch type is very useful to warn about type
promotion issues but it's not really correct.

This add a new helper for when we know that the type is a comparison.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>