ssa, unwind: instead of tracking the variable track the value
One of the common causes of false positives for the check_unwind.c check
is when code does this:

	ret = alloc_something();
	if (ret < 0)
		return ret;
	foo->something_id = ret;

	ret = alloc_something_else();

In the original code, if we freed a resource "release(foo->something_id);"
the the check looked for other names for the resource and marked those as
freed too.  But in this case "ret" has been reassigned to something else so
it's not equivalent at the end.  So it was marked as a leak.  Which is a
false positive.

What we need is something like Single Static Assignment (SSA) where we
don't track the variable we track the resource.

The first thing we need to do is give the resource a name.  In this case
we just have a counter to which gives out unique numbers. But we also want
a human readable name so the name is a combination of the first variable
and the unique number "ret:12354".  There is no symbol associated with the
SSA sm_state.

So to store a resource as allocated it's a two step process.  First get
the ssa name "ret:12354" then set "ret:12354" to allocated.  There is a
function set_ssa_state() which handles this.  If you already know the ssa
name then use update_ssa_state() which is just a wrapper around
set_state().

The set_ssa_state() is actually slightly more involved than that.  We only
allocate ssa states when they are needed.  If we have an unmatched state
then it could be that we have an &undefined state so we need to allocate a
new ssa state in that case.  I'm not sure the code to handle &merged states
is sufficient, but it isn't an issue yet...

It's probably better to use the get_ssa_sm_state() function instead of
the get_ssa_state() function.  (The difference is the former returns the
sm_state and the latter returns the smatch_state).  You want the sm_state
because then you can update the state without doing a second look up to
find the ssa name.

There is also the ssa_name() function which strips off the ":12354" part
of the name.  Probably that name is confusing given how I have been using
it to mean the reverse in this commit message.

The other thing that I did was that I made the "info" part of this where
we record which parameters are freed separate from checking for leaks.
The logic for these two pieces of code is slightly different and how
programmers treat parameters is slightly different so this works pretty
well.  I also made the unmatched_state() function static as it should have
been from square one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>