kernel_user_data: fix copy_from_user() again
Somehow the copy_from_user() function has started to be recognized as a
memset() which resets the destination buffer.  The problem is that these
are handled in the wrong order so it first gets set to user data, and
then reset back so it is no long marked as user data.  Which is totally
useless.

Fix this by saving the copy_from_user() expression and ignoring it when
it comes to marking the parameter as cleared.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>