hooks: add a new add_array_initialized_hook() hook
I had two use cases for this, but only one turned out to be useful so far.

It turns out that one thing which people do is create a struct with:

	int num_foo;
	struct my_struct foo[10];

And the "num_foo" store how many of the elements of "foo" are initialized.

This patch adds some new code so that you can record how many elements of
an array are initialized at declaration time.  The remaining elements will
be zeroed, of course.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>