locking: ignore efx_rwsem_assert_write_locked()
The efx_rwsem_assert_write_locked() function seems tricky to me to handle.
It tries to take a semaphore and if it's able to then it prints a warning
and releases the lock.

It's not recorded in the DB as a lock change, but somehow when it's parsed
inline then I guess it is...  I'm not sure what's going on with that.  Then
what happens is that it gets saved as a LOCK in the caller and it ends up
generating tons of double lock warnings.

It's sort of a weird thing because I would almost think that
efx_rwsem_assert_write_locked() should return a KNOWN_UNLOCKED because we
either fail to take the lock or we release it.  But that's the opposite
because at the end of the function we are ideally in a known locked state.
Except that's not really true either, the lock could be held by a different
thread.

The simplest fix is to just ignore the efx_rwsem_assert_write_locked()

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>