free_strict: silence some kfree_skb() false positives
Skb structs are reference counted so if we call skb_get() then calling
kfree_skb() doesn't actually free it, it just drops the reference count.

The main problem with this code is that it has the wrong name for the
reference counter.  It should have been "->users.refs.counter" instead of
"->users.counter".

There are couple other improvements, in the original code, it looked for
kfree_skb() to apply this rule, but there are a bunch of ways to free
skbs so this rule should just apply to the type and not the free function.
Also use the rule for FREED from the db rules as opposed to just hard
coded free functions.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>