allocation_funcs: add a lot more functions
I'm likely to re-write this code soon because it's legacy and garbage.
What this code does is it helps generate the kernel.allocation_funcs.

There were a bunch of allocation funcs missing.  I found a more complete
list in smatch_buf_size.c and copied it over.

But the whole approach is crap.  This shouldn't be a list that gets copied
from one .c file to the next.  It needs to be stored in a central
per project table.  But for now, this is an improvement.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>