db: add support for deleting and splitting returns manually
I never actually ended up using this feature, but it probably will be
useful in the future.  It's actually two features.  In
smatch_data/db/fixup_kernel.sh then we sometimes delete return states
and this fine.  But it doesn't work for inlines.  Of course, we do have
the option of disabling inline parsing for that function but that's maybe
heavy handed.  What we could do instead would be to delete a given return.

The other option is that Smatch tries to figure out how to split the
returns into something useful like NULL vs non-NULL.  But maybe sometimes
you want to manually control how returns are split and this gives you a
way to do that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>