param_set: setting stack data doesn't not affect the caller
This is an awkward patch...  The problem is this:

	foo = parameter->whatever;
	foo = something_else;

In the original code, what happened was the first line would say that
"foo is parameter->whatever".  Fine.  Then the next line it would say
"Ah, we're setting foo, but foo is parameter->whatever so we're setting
parameter->whatever".  Which is absolutely not fine.

So really this is a kind of ordering bug.

But instead of fixing the ordering bug, what I did was I said that if we're
setting a stack variable like "foo" (as opposed to "foo->bar") then that
can't affect outside the function.

But then it affected param_was_set_var_sym() because now if we have a
parameter:

void frob(struct foo *p)
{
	p = something_else();
	p->xxx = whatever;
}

Then "p" wasn't recorded as modified so param_was_set_var_sym() needed to
be updated as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>