kernel: delete PCIBIOS_BAD_VENDOR_ID returns
Some of these returns were already deleted via the fixup_kernel.sh script.
The motivation for that was because no one actually checks for errors.
(It's pointless to check, because if your PCIBIOS is not working then
software can't help you, you need to buy new hardware).

The reason I'm deleting them now is because there are too many error paths
so smatch stops parsing and doesn't see that these functions all set *data.
Smatch's limit for this is 100 return paths and deleting these takes us
from 123 return paths to 93 return paths.

Probably it won't take long to add more return paths and this solution will
break.  But whatever...  It works for now and we'll probably want this
feature in the future.  There isn't any benifit to keeping the PCIBIOS
error paths around.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>