math: assume container_of() return valid pointers
This code ignores container_of() when it is a no-op.  This is important
because sometimes the author knows that container_of() is a no-op so it
can have a NULL or an error pointer stored in it.

But if we take a pointer and subtract 16 from it, then we should be able to
assume that the result is a valid, non-NULL pointer.  If this is not true
then we are already in serious trouble and it's too late to do anything and
the assumption has not hurt us.

In other words there is no downside to this assumption and there is a good
upside because it lets us silence a bunch of warnings.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>