implied,db: fix a bug faking history
The problem is that in smatch db we go through all the leaf states to try
to separate the returns.  But when we fake smatch history an sm state
can be changed from a leaf state to a merged state.  It should still be
treated like a leaf state in smatch_db.c so add a flag which says that it
is still a leaf state in that case.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>