ssa: fix the modification hook
The modification hook was trying to work around a theoretical ordering
issue with the assignment hook but actually there isn't really an issue at
all.  We need to do the modification hook first (which is fine because
the modification hooks are already done very near the start) then checks
can set the state of expr->left, then in the assignment hook we can test
if the any checks set changed it from &undefined to something new.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>