ssa: introduce ssa_set_all_states()
Say we have:

	foo = bar;
	p->foo = foo;
	kfree(foo);

We want to mark both "foo", "bar" and "p->foo" as freed.  The traditional
way would be to open code something which looks at smatch_assignment.c
states.  But it's tricky because sometimes the foo could be on the either
the left side or the right side of the assignment.

It's cleaner both from a conceptual and code view point to do:

	ssa_set_all_states_expr(my_id, foo_expr, &freed);

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>