smatch.h: introduce sm_pedantic()
I created a --pedantic option which is for style issues and trivial stuff.
The problem is that I never build my kernel with the pedantic option
because it's too much work to sort through the warnings.  So it's better
to print them out but with the pedantic: prefix so that I can just use
grep for filtering.

I got rid of the ancient unused sm_spam() macro at the same time.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>