kernel_printf: print a warning for potential precision vs width
This check is only enabled with --spammy because it tends to have a some
amount of false positives.

Sometimes people use a precision specifier to print strings which are not
NUL terminated.  And sometimes they intend to do that but accidentally use
a width specifier instead.  Smatch is often able to track which strings
are NUL terminated but if it's not, and the name doesn't indicate it's
a width specifier then print a warning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>