db: prefer storing param data over return data
What this is is was we have a function that looks like:

struct whatever *my_function(struct whatever *p)
{
	frob(p);
	frob(p);
	return p;
}

In this function "p" is both a parameter and also the return value.  It's
more useful to talk about it in terms of a parameter instead of a return
value because the function might ignore the return, but it's always going
to pass a parameter.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>