smdb: add a "preempt" option
When I'm debugging a "sleeping in atomic" bug then I want to know which
functions call this function with a spinlock held.  This option prints
the preempt disabled call tree.  I changed the call_tree code to print
longer call_trees because some of the preempt call tress are very long.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>