validation: update locking tests
I basically re-wrote the locking check and the validation tests were out
of date.

The truth is there were some of the old warnings that I liked in theory
but it wasn't easy or worth it to implement in the new code.  So they're
gone.

I have some ideas for how to re-write the locking check again and I will
end up doing that.  But for now the locking check is basically working
okay.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>