function_hooks: don't call both DB and implied return hooks callback
One problem with the param/key API is that you register the same call
back for both the implied return and the DB return and it gets called
twice resulting in a "double lock" or whatever.  The implied return
should trump the DB return and only be called once.  So it needs to be
tracked in smatch_function_hooks.c

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>