db/fixup_kernel.sh: hack around device_add()
The device_add() function is really important and Smatch doesn't parse
it correctly because there are just too many states.  So what I've done
as a hack around for this is to just delete the PARAM_ADD states.  This
deletes 1422 states and makes the function manageable.

It's an ugly hack.  And unfortunately, since I did the hack the number of
states has increased again to the point where device_add() is still not
parsed correctly.  :(

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>