db/fixup_kernel.sh: delete kobject_init() failure path
The kobject_init() function doesn't fail in real life.  If it did then we
are screwed already so there is no benefit in trying to calculate all the
implications.

Unfortunately, it's a void function so we can't delete the error path based
on the return value and have to do this complicated thing instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>