kernel: ignore a lot more stuff that we don't care about
The kernel has a bunch of internal data which we don't really care about.

For example, with mutexes, Smatch doesn't care about the lockdep.  I only
care if the lock is set or not.  But locks are really common and there is
a ton of struct members needed to support lockdep.  Let's filter that data
out instead of storing it in the database.

The impetus for this change was because I wanted to get device_register()
small enough to the point where we had enough space to store both the
success and the failure path.

I made some other changes to how smatch_param_limit/filter so it didn't
count unused states.  (This should have made the DB larger).  And I made
some changes to smatch_power_of_two.c that should have made the DB smaller.
With these changes my database is 6GB smaller than yesterday.  The
device_add() function is stored nicely now but device_register() is still
broken.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>