math: re-write handling of offset_of()
The handling of offset_of() had a comment, "FIXME:  I don't really know
what I'm doing here." which was based on a true story.  I have re-written
it.  It's mostly copied from the Sparse implementation so hopefully it
should be correct.

One of the differences between Smatch and Sparse is that for Smatch I
don't want to evaluate() the code.  A sizeof() should remain a sizeof()
and not get turned into a EXPR_VALUE.  So that's why we can't use the
Sparse code as-is, but it's basically identical.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>