flow: ignore BUILD_BUG_ON()
In the kernel, the BUILD_BUG_ON() macro was recently modified and declared
to be a no return function.  What happens is that people use BUILD_BUG()
to stop the build when things cannot be determined at build time.

static inline my_func(int x)
{
	if (x == 1)
		...
	else if (x == 2)
		...
	else
		BUILD_BUG();
}

Clang complains that a bunch of variables are not initialized on the
BUILD_BUG() path.  But now that BUILD_BUG() on is a no return function
it's equivalent to calling BUG() and the uninitialized variable warnings
are silenced.

The problem is that it breaks Smatch because sometimes people do:

static inline my_func(int x)
{
	BUILD_BUG_ON(!__builtin_constant_p(x));
	...
}

Smatch knows what the value of "x" is and sometimes treats it as a
constant.  But it's hard to know if GCC will always treat it as a constant.
Instead Smatch says, "No, "x" isn't a constant so the my_func() function
will never return".  All the callers get messed up as well.

So for Smatch it's more useful to ignore BUILD_BUG_ON() than to treat it as
a call to BUG().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>