math: improve how handle_builtin_constant_p() is handled
The kernel recently made BUILD_BUG_ON() equivalent to BUG_ON() in terms
of flow analysis.  Obviously BUILD_BUG() macros only affect the build and
not runtime, but this is a kind of trick to make static analysis easier
for Clang.

Unfortunately, the Smatch doesn't parse the code the same way that GCC does
so it can't always determine what is build time constant.  The end result
is that now Smatch thinks that some functions just call BUG() and never
return.

I tried to make Smatch follow GCC more closely and this was the result.  It
turned out to a bit of a hopeless task and so I modified Smatch to ignore
the BUILD_BUG() macros.  But the following GCC more closely is still useful
so I'm going to commit this as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>