uninitialized: don't warn about missing error checks for function pointers
The most common reason for uninitialized variable warnings is that people
expect a function to initialize the variable and it does not set it on the
error path.  Some of these have "theoretical" error paths, but which only
static analysis cares about but programmers understand there is nothing
which can be done to recover if read() functions start failing.  I can
add these to the ignore list.

But for function pointers the ignore list doesn't support that.

Screw it.  If people can't be bothered to check for errors at all, then
just assume they know what they are doing and don't warn about potential
bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>