uninitialized: warn about "uninitialized_var |= foo;" type assignments
This check doesn't find any bugs, but there are other tools which have a
similar check so I assume that all the kernel bugs were fixed already.

The trick of this patch is that the ordering had to be done correctly
because extra_mod_hook() and match_assign() both set assignments to
initialized.  I'm a little bit confused by why the check in extra_mod_hook()
has && instead of || but it had to do with getting the ordering correct
I think.

Also I had to move the is_initialized() function earlier so that makes the
diff larger.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>