uninitialized: silence false positives with enum switch statements
One common cause of false positives is when:

switch (some_enum) {
  <All values handled.  No default case>
}

Ideally if you had cross function analysis and whatever then Smatch would
know the possible values and handle this correctly.  Unfortunately a lot
of code is too complicated.

The next most ideal situation would be if Smatch could determine the values
of some_enum based on the enum_declaration.  But right now it doesn't and
that's hard to implement.

So this patch just silences the warnings in the check itself.  If the
switch is based on an enum, and there is no default case then mark
everything which is sometimes initialized as always initialized.  On my
system this seems to silence about 3% of my warnings.

Reported-by: Jesse Brandeburg <jesse.brandeburg@intel.com>